chore: Update NDI SDK path and add target arch to bindgen
Updates the standard NDI SDK path for Linux to `/usr`. This is
done to ensure the build process can find the NDI SDK in more
standard locations.

Also, adds the target architecture as a clang argument to bindgen.
This ensures that the correct headers are generated for the target
architecture when cross-compiling.